v0.7

+++ CODE/INTERNAL.py
@@ -11,7 +11,6 @@ os.system('cls' if os.name=='nt' else 'clear')
 if consoleRunning and systemOS != 'Windows':
     consoleRows, consoleColumns = os.popen('stty size', 'r').read().split()
 elif consoleRunning and systemOS == 'Windows':
-
     h = windll.kernel32.GetStdHandle(-12)
     csbi = create_string_buffer(22)
     res = windll.kernel32.GetConsoleScreenBufferInfo(h, csbi)
@@ -30,25 +29,14 @@ if consoleRunning: matplotlib.use('Agg')
 #If consistentcy in the random generator is desired for comparisons
 if consistentSeed: np.random.seed(0)
 
-#Initialize the information object for training and testing
-reconMethod = 'CWM'
-featReconMethod = 'CWM'
-neighborWeightsPower = 2
-numNeighbors = 10
-filterType = 'Gaussian'
-featDistCutoff = 0.25
-resolution = 1
-info = Info(reconMethod, featReconMethod, neighborWeightsPower, numNeighbors, filterType, featDistCutoff, resolution, imageType)
+#If incremental data is not being used, then a model must be trained for each c value
+if incrementalTrainingMasks == False: bestCFromTrainingData=False
 
-#Set the number of available CPU threads, leave 2 free if possilbe
-num_threads = multiprocessing.cpu_count()
-if num_threads > numFreeThreads: num_threads -= numFreeThreads
-
-#Setup worker pool; 50% for object storage, 10% for redis metadata/lineage, half-of-rest for workers
-amount_RAM = int(psutil.virtual_memory().available*(percRAM/100))
+#Set the multiple for which the input data must be for network compatability
+depthFactor=2**numConvolutionLayers
 
 #Initialize multiprocessing pool server
-ray.init(num_cpus=num_threads, logging_level=logging.ERROR)
+ray.init(logging_level=logging.ERROR)
 
 #Print and store all current variables in RAM
 #variableList = []
@@ -69,7 +57,8 @@ dir_ImpData = dir_InputData + 'IMP' + os.path.sep
 #Results directories
 dir_Results = '.' + os.path.sep + 'RESULTS' + os.path.sep
 dir_TrainingResults = dir_Results + 'TRAIN' + os.path.sep
-dir_TrainingResultsImages = dir_TrainingResults + 'Images' + os.path.sep
+dir_TrainingModelResults = dir_TrainingResults + 'Model Training Images' + os.path.sep
+dir_TrainingResultsImages = dir_TrainingResults + 'Training Data Images' + os.path.sep
 dir_TestingResults = dir_Results + 'TEST' + os.path.sep
 dir_TestingResultsImages = dir_TestingResults + 'Images' + os.path.sep
 dir_ImpResults = dir_Results + 'IMP'+ os.path.sep
@@ -88,10 +77,15 @@ if not os.path.exists(dir_ImpData) and impModel: sys.exit('Error - dir_ImpData d
 if not os.path.exists(dir_Results): sys.exit('Error - dir_Results does not exist')
 
 #As needed, reset the results' directories
-if trainingModel:
+if trainingModel and not loadTrainingDataset:
     if os.path.exists(dir_TrainingResults): shutil.rmtree(dir_TrainingResults)
     os.makedirs(dir_TrainingResults)
-    os.makedirs(dir_TrainingResultsImages)    
+    os.makedirs(dir_TrainingModelResults)
+    os.makedirs(dir_TrainingResultsImages)
+
+if trainingModel and loadTrainingDataset:
+    if os.path.exists(dir_TrainingModelResults): shutil.rmtree(dir_TrainingModelResults)
+    os.makedirs(dir_TrainingModelResults)
     
 if testingModel:
     if os.path.exists(dir_TestingResults): shutil.rmtree(dir_TestingResults)
